Clear river ids at game start, store as short int.
Should generally store as little data as possible at CvPlot. River ids are only useful during map generation; shouldn't be stored at CvPlot at all. However, a few map scripts use river ids, so it wouldn't be easy to move that data out of CvPlot.

Prompted by this CFC post:
https://forums.civfanatics.com/threads/quick-modding-questions-thread.351519/page-290#post-15737032

+++ CvGameCoreDLL/CvMap.h
@@ -351,7 +351,7 @@ public: // advc: made several functions const
 	int getTopLatitude() const;																									// Exposed to Python
 	int getBottomLatitude() const;																							// Exposed to Python
 
-	int getNextRiverID() const;																									// Exposed to Python
+	short getNextRiverID() const;																									// Exposed to Python
 	void incrementNextRiverID();																					// Exposed to Python
 
 	bool isWrapXExternal(); // advc.inl: Exported through .def file							// Exposed to Python
@@ -485,7 +485,7 @@ protected:
 	int m_iOwnedPlots;
 	int m_iTopLatitude;
 	int m_iBottomLatitude;
-	int m_iNextRiverID;
+	short m_iNextRiverID; // advc.opt: was int (these get stored at CvPlot)
 
 	bool m_bWrapX;
 	bool m_bWrapY;
